+++ components/price-stabilization/DollarPrice.tsx
@@ -1,29 +1,12 @@
-import { useState, useEffect } from "react";
 import { ethers, BigNumber } from "ethers";
-import { parseEther } from "ethers/lib/utils";
 
 import { Tooltip } from "@/ui";
-import { useDeployedContracts, useManagerManaged, useWeb3Provider } from "../lib/hooks";
+import usePrices from "./lib/usePrices";
 
 const roundPrice = (twapPrice: BigNumber): string => parseFloat(ethers.utils.formatEther(twapPrice)).toFixed(8);
 
 const DollarPrice = () => {
-  const web3Provider = useWeb3Provider();
-  const deployedContracts = useDeployedContracts();
-  const managedContracts = useManagerManaged();
-  const [twapPrice, setTwapPrice] = useState<BigNumber | null>(null);
-  const [spotPrice, setSpotPrice] = useState<BigNumber | null>(null);
-
-  useEffect(() => {
-    if (web3Provider && managedContracts && deployedContracts) {
-      (async () => {
-        const newTwapPrice = await managedContracts.twapOracle.consult(await deployedContracts.manager.dollarTokenAddress());
-        const newSpotPrice = await managedContracts.metaPool["get_dy(int128,int128,uint256)"](0, 1, parseEther("1"));
-        setTwapPrice(newTwapPrice);
-        setSpotPrice(newSpotPrice);
-      })();
-    }
-  }, [web3Provider, managedContracts, deployedContracts]);
+  const [twapPrice, spotPrice] = usePrices();
 
   return (
     <div className="mb-4">
